separate and add app into Landing, Scanning, History and Education screens
