+++ frontend/App.js
@@ -1,48 +1,43 @@
 import React, { useState } from 'react';
 import { StyleSheet, Text, View, Button, TextInput } from 'react-native';
-import axios from 'axios';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+
+import LandingScreen from './screens/LandingScreen';
+import ScanningScreen from './screens/ScanningScreen';
+import HistoryScreen from './screens/HistoryScreen';
+import EducationScreen from './screens/EducationScreen';
+
+
+const Tab = createBottomTabNavigator();
 
 export default function App() {
-  const [response, setResponse] = useState('');
-  const [number, setNumber] = useState('');
-
-  const pingServer = async () => {
-    try {
-      const res = await axios.get(`http://100.69.247.33/ping/${number}`,);
-      setResponse(res.data);
-    } catch (error) {
-      console.error(error);
-      setResponse('Error pinging server');
-    }
-  };
 
   return (
-    <View style={styles.container}>
-      <TextInput
-        style={styles.input}
-        onChangeText={setNumber}
-        value={number}
-        placeholder="Enter a number"
-        keyboardType="numeric"
-      />
-      <Button title="Ping Server" onPress={pingServer} />
-      <Text>{response}</Text>
-    </View>
+    <NavigationContainer>
+      <Tab.Navigator initialRouteName="Home">
+        <Tab.Screen
+          name="Home"
+          component={LandingScreen}
+          options={{title: 'Welcome'}}
+        />
+        <Tab.Screen 
+          name="Profile" 
+          component={ScanningScreen} 
+          options={{title: 'Scanning'}}
+        />
+        <Tab.Screen 
+          name="History" 
+          component={HistoryScreen} 
+          options={{title: 'History and Results'}}
+        />
+        <Tab.Screen 
+          name="Education" 
+          component={EducationScreen} 
+          options={{title: 'Learn More!'}}
+        />
+      </Tab.Navigator>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  input: {
-    height: 40,
-    margin: 12,
-    borderWidth: 1,
-    padding: 10,
-    width: '80%',
-  },
-});
