+++ frontend/screens/ScanningScreen.js
@@ -1,0 +1,29 @@
+import React, { useState } from 'react';
+import { StyleSheet, Text, View, Button, TextInput } from 'react-native';
+
+export default function ScanningScreen({navigation}) {
+  
+    return (
+  
+      <View style={styles.container}>
+        <Text>Scanning</Text>
+      </View>
+    );
+  }
+  
+  const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    input: {
+      height: 40,
+      margin: 12,
+      borderWidth: 1,
+      padding: 10,
+      width: '80%',
+    },
+  });
+
